pyproject.toml
[project]
name = "fantasyfootballai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.1",
    "python-dotenv>=1.1.1",
]
